// fe/src/app/api/admin/model/pageResponseOrderResponse.ts
/**
 * Shop Admin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderResponse } from './orderResponse';


export interface PageResponseOrderResponse { 
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
    data?: Array<OrderResponse>;
}